發現一個可以在grades表新增的方法

+++ database/migrations/2015_05_06_003652_create_users_table.php
@@ -14,7 +14,7 @@ class CreateUsersTable extends Migration {
 	{
 		Schema::create('users', function(Blueprint $table)
 		{
-			$table->integer('id')->unique()->unsigned();
+			$table->bigInteger('id')->unique()->unsigned();
 			$table->string('name');
 			$table->string('password');
 			$table->string('phone')->default('');
